better handling of initial setup (don't redo it every mission)

+++ Script/libsystem.ks
@@ -2,14 +2,23 @@
 //print "  Loading libsystem".
 
 function loadPersistent {
-    log "" to "1:/persistent.ks".
-    RunPath("1:/persistent.ks").
+    if Exists("1:/persistent.ks") RunPath("1:/persistent.ks").
 }
 
 function writePersistent {
-    log "" to "1:/persistent.ks".
-    DeletePath("1:/persistent.ks").
+    if Exists("1:/persistent.ks") DeletePath("1:/persistent.ks").
     log "set pMissionCounter to " +pMissionCounter +"." to "1:/persistent.ks".
+}
+
+Function askConfirmation {
+  parameter deadline is Time:Seconds+1e9.
+  set AG1 to false.
+  print "  == press AG1 to continue ==".
+  until (AG1 or (Time:Seconds>deadline)) {
+    print "dt =" +Round(deadline-Time:Seconds)+"  " at (38,1).
+    //print "AG1=" +AG1+"  " at (38,1).
+    wait 0.
+  }
 }
 
 function setControlPart {
@@ -18,7 +27,8 @@ function setControlPart {
     pList[0]:ControlFrom.
   else
   {
-    print "ERROR: no unique control part found: num="+pList:Length.
-    interruptMission().
+    print "no unique control part found: num="+pList:Length.
+    print "please activate control part manually".
+    askConfirmation().
   }
 }
